+++ include/tulips/transport/ena/Device.h
@@ -41,7 +41,7 @@ public:
   Status wait(Processor& proc, const uint64_t ns) override;
 
   Status prepare(uint8_t*& buf) override;
-  Status commit(const uint32_t len, uint8_t* const buf,
+  Status commit(const uint16_t len, uint8_t* const buf,
                 const uint16_t mss) override;
   Status release(uint8_t* const buf) override;
